[PjRt-IFRT] Add optional global device mapping support to PjRt-IFRT
This change adds a capability to PjRt-IFRT to optionally take a user-specified
global device mapping. PjRt-IFRT will materialize global (multi-host) view for
IFRT devices even if the wrapped PjRt client may only provide a local
(single-host) view with addressable devices only.

This capability is achieved by forging `xla::GlobalTopologyProto` from the
device information in the provided `pjrt_client` and the global device mapping
information in `xla::ifrt::PjRtClient::CreateOptions::global_device_mapping`.

The current iteration of the global view (when created from
`global_device_mapping` and not obtained from `pjrt_client`) enables replicated
execution of XLA computations. SPMD executions are not yet supported.

This global device mapping feature is available when not using topology
exchange via a key-value store. The global topology information from local
topology exchange via the key-value store is used as-is by PjRt-IFRT as before
at the moment.

A side effect is that PjRt-IFRT devices have a slightly different format for
`ToString()` and `DebugString()` if device ID is remapped. The new device ID
will be noted as `"[PjRtIFRTDeviceId=XX]"` in the end of the device debug
string; if the device ID is the same between PjRt and PjRt-IFRT, the device
debug string remains the same. This small change intends to preserve the
backward compatibility with most user code that does pattern matching on the
device debug string (though the user code is discouraged to do so).

This change adds a two-process version of the test library using PjRt CPU
clients, and use it for PjRt-IFRT implementation tests as a simple
demonstration of the capability.

PiperOrigin-RevId: 698582589